Add default config vars for dev and move stuff to common
