+++ frontstage/common/eq_payload.py
